Kode/requirements.txt
# Core dependencies for groundwater risk assessment workflow
# Install with: pip install -r requirements.txt

# GIS and spatial analysis
geopandas>=0.12.0
shapely>=2.0.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Optional: for extended analysis
# openpyxl>=3.1.0  # Excel file formatting (optional_analysis only)
# folium>=0.14.0   # Interactive maps (optional_analysis only)
